requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
google-generativeai
faiss-cpu
pypdf
python-dotenv
supabase
sentence-transformers
email-validatorsendgrid